rename test/unit/misc-assets to test/unit/assets
