+++ src/render/glyph_manager.test.ts
@@ -4,7 +4,7 @@ import fs from 'fs';
 import {RequestManager} from '../util/request_manager';
 
 const glyphs = {};
-for (const glyph of parseGlyphPBF(fs.readFileSync('./test/unit/misc-assets/0-255.pbf'))) {
+for (const glyph of parseGlyphPBF(fs.readFileSync('./test/unit/assets/0-255.pbf'))) {
     glyphs[glyph.id] = glyph;
 }
