+++ test/unit/assets/mbsv5-6-18-23.vector.pbf
